requirements.txt
# Core data build scripts
beautifulsoup4>=4.12.0

# Google Sheets inventory sync
gspread>=6.0.0
google-auth>=2.0.0

# Scraping and automation tools used in Scrape/
requests>=2.31.0
selenium>=4.20.0
networkx>=3.0
pyvis>=0.3.2
